Move the dethmapX views into separate folders
