+++ GuiUnitTest/testviewhelpers.cpp
@@ -13,7 +13,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-#include "viewhelpers.h"
+#include "depthmapX/views/viewhelpers.h"
 #include "catch.hpp"
 #include <time.h>
 #include <sstream>
