+++ depthmapX/views/glview/gldynamicline.cpp
