+++ depthmapX/views/glview/gldynamicrect.cpp
