+++ depthmapX/views/glview/gldynamicrect.h
