+++ depthmapX/views/glview/glpointmap.cpp
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-#include "depthmapX/glpointmap.h"
+#include "glpointmap.h"
 #include "salalib/linkutils.h"
 #include "salalib/geometrygenerators.h"
