+++ depthmapX/views/glview/glpolygons.cpp
