+++ depthmapX/views/glview/glrastertexture.cpp
