+++ depthmapX/views/glview/glshapegraph.cpp
