+++ depthmapX/views/glview/glshapemap.cpp
