+++ depthmapX/views/glview/gltrianglesuniform.cpp
