+++ depthmapX/views/glview/glutriangulator.cpp
