+++ depthmapX/views/mapview.cpp
