+++ depthmapX/views/plotview/plotview.cpp
