+++ depthmapX/views/tableview/tableview.cpp
