+++ depthmapX/views/viewhelpers.cpp
