+++ depthmapX/views/views.pri
@@ -1,0 +1,37 @@
+HEADERS += $$PWD/mapview.h \
+    $$PWD/viewhelpers.h \
+    $$PWD/glview/gllinesuniform.h \
+    $$PWD/glview/glview.h \
+    $$PWD/glview/gllines.h \
+    $$PWD/glview/glrastertexture.h \
+    $$PWD/glview/glpolygons.h \
+    $$PWD/glview/glutriangulator.h \
+    $$PWD/glview/gltrianglesuniform.h \
+    $$PWD/glview/glpointmap.h \
+    $$PWD/glview/glshapegraph.h \
+    $$PWD/glview/glshapemap.h \
+    $$PWD/glview/gldynamicrect.h \
+    $$PWD/glview/gldynamicline.h \
+    $$PWD/plotview/plotview.h \
+    $$PWD/tableview/tableview.h \
+    $$PWD/depthmapview/depthmapview.h \
+    $$PWD/3dview/3dview.h
+
+SOURCES += $$PWD/mapview.cpp \
+    $$PWD/viewhelpers.cpp \
+    $$PWD/glview/gllinesuniform.cpp \
+    $$PWD/glview/glview.cpp \
+    $$PWD/glview/gllines.cpp \
+    $$PWD/glview/glrastertexture.cpp \
+    $$PWD/glview/glpolygons.cpp \
+    $$PWD/glview/glutriangulator.cpp \
+    $$PWD/glview/gltrianglesuniform.cpp \
+    $$PWD/glview/glpointmap.cpp \
+    $$PWD/glview/glshapegraph.cpp \
+    $$PWD/glview/glshapemap.cpp \
+    $$PWD/glview/gldynamicrect.cpp \
+    $$PWD/glview/gldynamicline.cpp \
+    $$PWD/plotview/plotview.cpp \
+    $$PWD/tableview/tableview.cpp \
+    $$PWD/depthmapview/depthmapview.cpp \
+    $$PWD/3dview/3dview.cpp
